backend/requirements.txt
fastapi
uvicorn
pyzbar
Pillow
google-generativeai
python-multipart
pydantic
ultralytics
numpy
opencv-python-headless
supabase==2.15.3
python-dotenv
